Updated openstack/openstack Project: openstack-infra/project-config  8dd71f39294412096086c1c43584434cd0fc3a6e
Speed up release-critical post jobs

To finalize the release process for a deliverable, we publish an
announcement pointing to the releases website (docs.o.o/releases).

We therefore need the openstack/releases publish-release-history
job to complete relatively quickly. However with the current
precedence=low on the "post" pipeline, this usually takes several
hours to complete, making this (by far) the longest step in the
otherwise-fast release process.

This change introduces a specific "release-post" pipeline with
precedence=high to which release-process-critical post jobs can
be assigned. Those jobs are rare enough (and complete fast enough)
so that the impact of that higher precedence on other jobs is
limited.

Change-Id: I2cf6355c4dd41dfd53f75cc4d0238f4b45b8eb9a