Updated openstack/openstack Project: openstack/ironic-specs  5c37955a4bc75f3eb2adbd403d60706a3b969829
Add notification support to Ironic

Ironic has a need for inter-service notifications. Near real-time
inter-service notifications in OpenStack have a variety of use cases.
Some of the use cases discussed for ironic included integration with
billing systems, using notifications to support an operations panel to
keep track of potential failures, and integration with TripleO. This
specification outlines a general framework for a notification schema in
ironic.

Change-Id: I67b1e2a815cd865792fba44362d94d272f69efcd