Updated openstack/openstack Project: openstack-infra/project-config  213913acada17fae9705a3fa555543bf7ee46701
puppet: pin bundler to 1.10.6

Puppet OpenStack CI is hitting this bug:
https://github.com/bundler/bundler/issues/4149

It affects the latest bundler release that is used in Puppet OpenStack
CI jobs (and breaking syntax3 and unit3.x).

The patch that fix bundler has been fixed but is not in the latest
release.
We had 2 options to fix it:
* patch 26 modules with the dependency in Gemfile, plus backport patches
  in stable branches, which is around 100 patches + 100 other patches
  when bundler is fixed (200 patches).
* patch jenkins/jobs/puppet-module-jobs.yaml to pin bundler with a
  previous release, until bundler has a new release.

Puppet OpenStack group decided to choose option #2 and fix it by pinning
Bundler to 1.10.6.

Bundler team does not seems ready to provide a release:
https://github.com/bundler/bundler/issues/4149#issuecomment-164527861

Change-Id: I907c595be0e7483eb0c3db545e36bf036d601b6d
Closes-Bug: #1525929