Updated openstack/openstack Project: openstack/tempest  9efd4c1bd77077c7279a5572e15211a19f721af4
Rename add_image_member to create_image_member

[POST /resources] methods should be "create_<resource name>",
so this patch applies the rule to v2 images_client.

To be honest, I felt add_image_member was not so wrong.
However, API doc also mentions "Create image member" as the API
and it is nice to rename it for the consistency of API doc and
the other APIs.

Partially implements blueprint consistent-service-method-names

Change-Id: I4f9eb9f65557197fdf54100a26531a63d6c1dcce