Updated openstack/openstack Project: openstack/taskflow  cc27e44751c8e3803467dad2adb4edce2fbc00fa
Change engine 'self._check' into a decorator

Instead of having 'self._check' be called at the stat
of every using function it can be made into a decorator
and placed on the function instead. This way it is a little
more explicit than having it called at the start of each
method.

This also adds a check to ensure the engine has been validated
before it is ran (and the decorator takes into this into
account as well).

Change-Id: I67ee8fe8629392003f7777d4d4fab918b1a071dd