Updated openstack/openstack Project: openstack/nova  050e691948bcab9dfdd6f8707700f9bb31464fee
Allow to update 'v2.1' links in sample files

Currently logic in _update_links() will update the 'v2'
versions in sample files links, we should have that logic
to update both 'v2.1' and 'v2' versions in links.

If there will be 'v2.1' versions in links, current logic
will find 'v2' in 'v2.1' text and replace it with 'v2.1'
and 'v2.1.1' for v2 and v2.1 scenario tests respectively.

This is needed to makes sample files in doc/api_samples
with 'v2.1' in links.

Change-Id: Id2d37520cc72bc5539b5f000b80c301d1fa7e3f5