Updated openstack/openstack Project: openstack/neutron  a25eb0657f3d99fa93e73245e18d9bec99219bf4
Removing adv svcs dependencies on neutron

After upstreaming Idcfeee6ddc2982fff49f44064c918bfd4627be12 and
I62bfd6b34c381427425ca2a656e5782087de3a3b, we can remove items
from Neutron.

Specifically, LOADBALANCER_PLUGIN and LOADBALANCER_AGENT are not
used now. FIREWALL_PLUGIN is used in one module, so moved there.
The LOADBALANCER constant is only used in test files, but it is
used in multiple files, so leaving it. LOADBALANCERV2 is only used
in the header it is defined, so leaving it as well.

In the FW and LB repos, there are no longer dependency on these literals.

Change-Id: I7f0b972e1ae7ef435a3612e239dad7b2c04baea2
Implements: blueprint neutron-lib