Updated openstack/openstack Project: openstack-infra/storyboard  35b19f9cbee81a4baa812bfe4bb0d13c47f5b2f2
Parse path correctly when it begins with /api

The regex for parsing request paths doesn't account for the fact that
the production API is at `$fqdn/api/v1/`. This causes the notification
system to not work entirely as expected.

This patch changes the regex to allow the path to begin with `/api`.

Change-Id: I846b8cad6b807c9350f8af85f7e68c870a31286a