Updated openstack/openstack Project: openstack-infra/shade  301e99ba9747c0cf080e9065bdf7f9ca359c227c
Fix server deletes when cinder isn't available

This commit fixes an issue when trying to delete a server and cinder
isn't available. Cinder isn't actually a required component of a cloud
and some public clouds (like vexxhost) don't deploy it, so we shouldn't
unconditionally attempt to talk to cinder without being able to handle
this case.

Change-Id: Ia1e19aad711c43fec21e2c1e5ad64b69acb561d3