Updated openstack/openstack Project: openstack-infra/puppet-kibana  df57bd6f9edf7d839cd40fdd3e7bb8b60bcfd35e
Encode query uri component properly

Previously we used encodeURI() to encode a single component of a URI,
this meant that characters like = were not encoded. This isn't correct
bceause we are just encoding a single component and need = to be
encoded. Fix this by using encodeURIComponent which should do the right
thing.

Change-Id: Ib320c1e646ccbd2233bd9d7a2be9df29ff7bd1ed