Updated openstack/openstack Project: openstack/nova  150d7b12654bf4110492bd66434331680f3b010f
rework warning messages for extension whitelist/blacklist

The API v2.1 extension loader was issuing a WARN for every single
extension that was listed in a white or blacklist. This is not a
warning for the administrator, this is a thing they chose to do. We
already have an INFO message that says all the extensions that got loaded.

The only WARN we should consider is when they put items in both the
white and black lists, and we decided to ignore the whitelist. That
warning message is added, as that's one of those things that could
drive people nuts trying to figure out why something didn't load when
added to the whitelist.

The function is denormalized and renamed to be more explicit about
what's going on.

Change-Id: I3ee8f04867509072c7cd6c0f1c1a87f893706595