Updated openstack/openstack Project: openstack/governance  0e0950db50e44fe186bf47b296f38551c32296b2
Redefine Release Management team mission

With the spin-off of the Stable branch maintenance subteam into
its own independent team, it makes sense to revisit the mission
of the Release Management team.

The old "Release Cycle Management" team used to regroup the
Release team, the Stable branch team and the Vulnerability
Management team. Now it's only the Release team left, so we
should also rename it to the name that everyone is calling us
anyway ("Release Management").

Also update the IRC channel and the wiki landing page.

Change-Id: I71b77d350dea2d2e697e6c59792d38a0274bda28