Updated openstack/openstack Project: openstack/nova  b9547181b1e550122431709fb5819174ca26d362
Ironic: Workaround to mitigate bug #1341420

The bug #1341420 can cause the driver to try to deploy onto a node which
already has an instance associated with it, the way Ironic reserve a
node to a specific instance is associating the UUID of that instance
with the node via the instance_uuid field. This patch just makes the
request to associate that field non-retriable when the api returns HTTP
409 (Conflict) (that's the HTTP error indicating that the node is
already associated with another instance), so it can fail fast in
case the node is already pick and the RetryFilter could then pick
another node for that instance.

Partial-Bug: #1341420
Change-Id: I697c5129893b18562182bbb3ceb8cb160e84c312