Updated openstack/openstack Project: openstack/neutron-fwaas  efb10ee9af743fd4d5098927fe19396b85a674aa
Setup for translation

To start translation, we need to initially import the translation
file - and place it at the proper place so that the usual CI
scripts can handle it. The proper place is for all python
projects $PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports will be done by the OpenStack Proposal bot.

Change-Id: Ica2b1f8adc611af5cfa2cef715526d2764c62427