Updated openstack/openstack Project: openstack/neutron-vpnaas  bc791b4bc2809bbbdfaf51d7c8db716c9378ce14
Setup for translation

To start translation, we need to initially import the translation
file - and place it at the proper place so that the usual CI
scripts can handle it. The proper place is for all python
projects $PROJECT/locale/$PROJECT.pot - see setup.cfg.

Further imports will be done by the OpenStack Proposal bot.

Change-Id: Ic405c8a808ee4eae0393e6f08da8f1569b326a59