Updated openstack/openstack Project: openstack/ironic  cc22b1916875adf9603f52b1aeebf97be7dda8de
Add bashate tox command

Now that we copied the devstack code to our tree we need to ensure that
the bash scripts we have are correct. This patch is adding a tox command
(bashate) that will check the syntax of the shell scripts under
the devstack/ directory.

The scripts under devstack/tools/ are not being checked, that's because
they weren't check when the code was in devstack as well and there's
loads to fix. It will be fixed in a following patch.

Change-Id: I8f66492081e0fd10e1de057febb9324a81411c7a
Partial-Bug: #1526518