Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  097a70ca9203d51fd1604b0d11fd7c534c0bc5ed
publishers: iterate over copy of checkstyle data

Iterate over a copy of checkstyle's data in both Python 2.x and Python
3.x. Python 3.5 does not allow us to modify the data OrderedDict during
a loop.

Note that this patch addresses the following error message which
is new in Python 3.5. That is, you need to be using Python 3.5+
to reproduce this error.

    RuntimeError: OrderedDict mutated during iteration

This bug is described at
https://storyboard.openstack.org/#!/story/2000436

Change-Id: Ia000279c38b991f4194cb587160b7a174872d991