Updated openstack/openstack Project: openstack/neutron  c7163fb7fdb3d9bf790ec27353ad4f46f2e93f84
Use a joined relationship for AZ info on networks

The previous code was doing a DB lookup for each network's
availability zone which was significantly impacting the
performance of network listings.

This patch adjusts the network model to be automatically joined
to the DHCP agents table that the AZ code uses to populate the
AZs for the network.

Change-Id: I908ceb1a68e0eed7c304e3ff82279ad6fa406167
Closes-Bug: #1525740