Updated openstack/openstack Project: openstack/ceilometer  deb8685e4db6163f48e56d74cc8c247437819188
verify gnocchi connection before processing

currently we create a client on initialisation but never verify that
the client has a connection. this creates a lot of errors in logs
because gnocchi isn't started yet.

this patch verfies the connection before it begins processing data by
calling gnocchi on initialisation.

DocImpact
Closes-Bug: #1489949
Change-Id: Ie3b7d966f068611e3fb60089b724f1a4c02f771e