Updated openstack/openstack Project: openstack/neutron  ab8adcf69380d1042c3440fe40b8a886fb855f45
Tune _get_candidates for faster scheduling in dvr

Right now we have seen some performance issues when
dvr routers are scheduled on multiple compute nodes
with thousands of VMs on the routed subnets.

The _get_candidates call get_l3_agent_candidates with
a complete list of agents irrespective of the routers
already hosted on the agents or not.

So this fix will reduce the amount of iterations that
get_l3_agent_candidates need to process for all the
agents and would increase the control plane performance.

Closes-Bug: #1513678
Change-Id: I8f781d4cbc996ce13441303c9296e4f6ec822b94