Updated openstack/openstack Project: openstack-infra/devstack-gate  e883f77282b46e32c3206bb52eb5d7b9079f15f2
Bump ironic timeout way up

Ironic builds on slower hosts tend to time out while PXE booting the
deploy ramdisk while transferring the image, producing great havoc on
the gate. We're investigating various approaches to fixing this (for
example, switching kvm's network driver didn't seem to help much), but
in the meantime let's be good citizens of gate resources and let things
run in full, even if PXE booting is slow.

This bumps the ironic timeout to 600, which should be more than enough
time for pxe booting to happen on slow hosts. It bumps both the value
that Tempest needs and the value ironic uses to time out a deploy.

As tempest defaults to 500 seconds for OS_TEST_TIMEOUT in .testr.conf,
we also need to bump that. To do so, this refactors the code that
launches tox to be able to inject OS_TEST_TIMEOUT, if it's set. We need
to inject that in the command directly, since tox is launched with sudo
and environment variables are not carried into that process.

Change-Id: I79c2b026b3f0bac0b96edef0418546bc4e480e3c
Related-Bug: #1393099