Updated openstack/openstack Project: openstack/python-novaclient  88ef0542dc3db5b9b98ae004087e90c123f68ece
Migrate to keystoneauth from keystoneclient

As a stepping stone to the os-client-config patch, first switch to
using keystoneauth, its Session and its argparse registration and
plugin loading to sort out any issues with that level of plumbing.
The next patch will layer on the ability to use os-client-config
for argument processing and client construction.

Change-Id: Id681e5eb56b47d06000620f7c92c9b0c5f8d4408