Updated openstack/openstack Project: openstack/nova  89dbc207c2b6cf641d9b2cd1e5243d7a9ce764c0
Add a note about fixing "db type could not be determined" with py34

Due to testrepository bug 1212909, if you run py27 before py34, py34
fails with "db type could not be determined". The workaround is to
delete the .testrepository directory and try again. As long as you run
py34 before py27 you're OK.

Change-Id: I13c4f9a39b97573e78fe397efc7d87f78bac4a58