Updated openstack/openstack Project: openstack/nova  668e371ed5a98c1055f2198d7af66b5f8c5fe4b9
Remove Instance object flavor helper methods only used in tests

These methods used to help with on-the-fly conversion of stored flavor
info from it's old location in system metadata to instance extra. Since
this is no longer required and since the only remaning uses are in
tests, remove the methods and fix the tests as needed.

Change-Id: Ia96106602be5cb256f78e3fe71e908627120657a