Updated openstack/openstack Project: openstack-infra/irc-meetings  0e5f3dfc7595aa10d7c9289d6c4d26a1fc1d5e70
Schedule the stable maintainers team meeting

As discussed in the openstack-dev ML [1].

The days of the week and times are mostly random, just looking for times
when I can reasonably expect to chair and trying to get some overlap with
Australian and European timezones.

The Monday meeting is the opposite week of the Neutron team meeting at the
same time.

[1] http://lists.openstack.org/pipermail/openstack-dev/2015-December/082464.html

Change-Id: Ib26be5f63ba6117a0bd332b5a4b7d0d827411318