Updated openstack/openstack Project: openstack/ironic  8eb1201145f07564432efae9298a22767d33c30a
Add grenade plugin

This is a straight copy of the ironic code in grenade's tree to here.
This is necessary to drop ironic code from devstack and complete the
devstack plugin work, as well as just the right thing to do, and will
help us move faster on grenade work.

Note that I'm not sure if this will or will not, and it may make our
grenade jobs explode during the transition. However, our grenade jobs
don't work right now anyway, so nothing of value was lost. :)

Change-Id: I8b065c60e9e06cd7fc10749c49bdf79d878fcc12