Updated openstack/openstack Project: openstack/nova  236bf85b2eaa19ab132ba1dad7be9d91f2dc3898
Add placeholders for config options

As the effort to centralize the config options progresses, we have found
that each commit that modifies the nova/conf/__init__.py script causes
merge conflicts for everyone else. In order to keep conflicts to a
minimum, we are adding lines for each of the eventual files, commented
out. As new sections are completed, the author simply has to uncomment
the relevant lines.

The file names are based either on existing grouping names, or on the
directory in which they were originally declared for those in the
DEFAULT grouping. There is some room for improvement, but this is a good
place to start arguing.

Blueprint centralize-config-options

Change-Id: I5a21afff87645244898c223b0022fa83978b1c7b