Updated openstack/openstack Project: openstack/horizon  54495c3185cae4920a56fa56d02768f893ee41f4
Allow specifying item to use for actions in the actions directive

Currently, the item passed to the 'allowed-actions' is aware of
what the HTML referred to. This patch allows for passing in the
item to use in the actions directive. This removes
the need to pass in the item key as part of the
allowed actions list thus removing the implicit binding
of the controller/service/html.

This also changes the list type attribute for the directive
from the action-list-type to type for readability.

Co-Authored-By: Kyle Olivo<keolivo@thoughtworks.com>

Partially-Implements: blueprint angularize-images-table
Change-Id: Idc344153aa7582c485a5d9c6b3cccbf8ff788d7c