Updated openstack/openstack Project: openstack/ceilometer  aa62b0dbf8c14d385fe3a8c31224ea11bebb5dd5
recheck cache after acquired gnocchi_resource_lock

Rechecking the cache after a sample got the gnocchi_resource_lock
because there is a possibility that the resource was already built
in cache by another ceilometer-collector.

This will reduce the unnecessay creation and update of resource in
gnocchi and improve the efficiency.

Change-Id: I85b30e770899c14ca06ae69b1f879106ceab0c5f
Closes-Bug:#1523193