Updated openstack/openstack Project: openstack-infra/project-config  4c5e35e2febab108f79d1e99d232f997524cadf1
Creation of the Coda Repo

At the Tokyo Summit, HPE asked if OSOps would take the responsibility to
house the Coda project. This is the creation of the repo for them to
upload the project to.

A description of Coda is, per Nathaniel Kimball:

Coda is a Horizon dashboard and panel (both share the name) that
facilitates resource clean up of a project once that project is no longer
needed (e.g. A customer cancels their account). It consists of a single
text entry box which allows you to search by project ID and will display
all resources, project info, and user info for the project. Upon pressing
the delete button the user is prompted each time for valid admin
credentials (annoying but a purposeful safe-guard) and will then walk the
list of resources deleting each one and will display the outcome of each
operation in a tabular format. This project started as a stand-alone
Django app inside HP and has been very valuable in operating our public
cloud. It was originally called Scrubber (because that’s what it did) but
was renamed to Coda after being ported to Horizon and adapted for Keystone
instead of a proprietary Identity service. It is the intent of HP to open
source our entire suite of tools using a musical theme. Coda was chosen as
the name for this project as a coda in music is a passage that brings a
piece to an end. It is particularly apropos for this use because the
symbol for a coda resembles a cross hair.

Change-Id: I08e550dd91575d2469cc6659a96002d786f2d372