Updated openstack/openstack Project: openstack/nova  808b84a43c4f0bd5c4ce2f1394e650c57cbed0b7
Add instance.save() when handling reboot in init instance

We have added a logic to handle nova-compute failure during
a soft reboot in patch: https://review.openstack.org/#/c/219980
but we have missed to add instance.save() to save the
modified task state, so the logic doesn't work.

This patch adds instance.save() at the end of this handling.

Change-Id: I0490643c5a1352ff13fd71d43be52260230f7df8
Closes-Bug: #1526162