Updated openstack/openstack Project: openstack/nova  2ba15705ab65211b1bf97ec152b720bf990e0a01
Move the Migration set-status-if-exists pattern to a method

We already use this pattern in places where we need to keep backwards
compatibility with an older RPC client that may not be sending the
migration object over.

Instead of it being a repetitive check, we make it a static
method on the ComputeManager class, and document that it can be removed
with the next major compute RPC version bump, in a single place instead
of a random comment.

Change-Id: I46632322f7abf3e329ebaab6587c7421d69bb861