Updated openstack/openstack Project: openstack/os-cloud-config  940d196c50c98a5bc156be9493661e18d3941886
Allow control over the Keystone adminurl and internalurl endpoint

We publish the keystone /v2.0 api in the endpoints by default and the
keystone v2.0 client will use the adminurl by default, not the publicurl,
so we need to allow a more granular configuration of the adminurl.

We also want the internalurl to stay on an internal network which is not
necessarily the same of either the public or the admin network so we need
to allow customization of the internalurl as well.

Change-Id: I6d6ad06441514c6eca9922937dba9b521dad06db