Updated openstack/openstack Project: openstack/nova  94200a682954327ae07e2ba2528680e1b48b6d37
Refresh stale volume BDMs in terminate_connection

Change Ib60d60a5af35be89ad8afbcf44fcffe0b0ce2876 meant
to address a race in deleting an instance while booting from
volume such that the BDMs passed to terminate_instance from
the API were stale (no volume_id set yet in the database).

Without the volume_id, _shutdown_instance can't detach the
volumes and _cleanup_volumes can't delete them.

The change fixed the race in _shutdown_instance by refreshing
the BDM list if necessary, but it didn't fix the problem of
deleting the volumes in _cleanup_volumes, which is still
getting a stale list of BDMs from the compute API.

This moves the refresh BDM logic to the top-level method
terminate_instance so that the refreshed BDMs are passed
to both _shutdown_instance and _cleanup_volumes.

There is some cleanup that can be done with
_delete_instance but since we'd like to backport this change
and the previous one, the cleanup is left for a later patch.

Change-Id: Ib0980a92ebdfa9c83496cd29bfea47fc42d538c5
Closes-Bug: #1464259