Updated openstack/openstack Project: openstack/trove  6490a6fed079bb53f166f02be72f03e9c49ca801
Modifying Vertica clusters to have a master node

In order to add management console functionality to Vertica
a master node is needed in the Vertica cluster. Changes are
made to rename the type of one instances in the cluster as
the master.

install_cluster and root enable api are directed to go to the
master node of the cluster.

Change-Id: I3c24531ce97e07d4b322849d7400d490ef8bd2e9
Closes-bug: #1523686