Updated openstack/openstack Project: openstack/tempest  3a0d3d081faba67b84eedf5491a6edf11b011517
Set a tempest lock_path default to a tmpdir

This commit sets a default value for the tempest lock_path to be just
be tempset-lock in the tempdir. Unlike the servers which could have
security implications for where the lock path is put, for tempest this
isn't a concern since it has to be manually invoked and has a finite
run time.

Change-Id: I2b326731984be8c9fd2491b52ffb1dfbc379f917