Updated openstack/openstack Project: openstack-infra/jenkins-job-builder  888928ee18ec3fd5e42a5f13eca4ec53e277286e
Infer and build entry points from component modules

Use the component type entry points to infer modules to scan
corresponding modules and automatically construct entry points for any
functions associated with yaml entries in JJB definitions.

This removes the need for new JJB modules added directly to the project
to also need to be listed in the setup.cfg to be reachable. Simply
adding the needed functionality to the correct file ensures that it is
available.

Change-Id: I2a0fd8f6720ddfd8582b58e47556a87ed1ad553d